Criado INSTALL.md e corrigido problema de cp sem ter bd.txt

+++ run.sh
@@ -11,7 +11,9 @@ chmod -R u=rwx,g-rwx,o-rwx $HOME/.web-server-pass
 cp send_email.py $HOME/.send_email.py
 chmod u=rwx,g=x,o-rwx $HOME/.send_email.py
 
-cp .database.txt $HOME/.database.txt
+if [ -f .database.txt ]; then
+	cp .database.txt $HOME/.database.txt
+fi
 
 ./passthrough -omodules=subdir,subdir=$HOME -o default_permissions -o allow_other -o auto_unmount -f mnt/
